src/requirements.txt
elasticsearch==9.3.0
pandas==3.0.2
python-dotenv==1.2.2
